refactor: regroup all BuilderDesign related files into design dir
