+++ frontend/src/features/admin-form-builder/common/components/BuilderSidebar/BuilderSidebar.tsx
@@ -4,7 +4,10 @@ import { Stack } from '@chakra-ui/react'
 import { BxsColorFill } from '~assets/icons/BxsColorFill'
 import { BxsWidget } from '~assets/icons/BxsWidget'
 
-import { DrawerTabs, useBuilderDrawer } from '../../BuilderDrawerContext'
+import {
+  DrawerTabs,
+  useBuilderDrawer,
+} from '~features/admin-form-builder/BuilderDrawerContext'
 
 import { DrawerTabIcon } from './DrawerTabIcon'
