+++ frontend/src/features/admin-form-builder/common/components/BuilderSidebar/DrawerTabIcon.tsx
