+++ frontend/src/features/admin-form-builder/common/components/BuilderSidebar/index.ts
