--- frontend/src/features/admin-form-builder/components/BuilderDesign/types.ts
@@ -1,9 +1,0 @@
-export type DndPlaceholderProps =
-  | {
-      droppableId: string
-      clientHeight: number
-      clientWidth: string
-      clientY: number
-      clientX: number
-    }
-  | Record<string, never>
