--- frontend/src/features/admin-form-builder/components/BuilderDrawer/CreateFieldDrawer/index.ts
@@ -1,7 +1,0 @@
-export {
-  CREATE_FIELD_DROP_ID,
-  CREATE_FIELD_FIELDS_ORDERED,
-  CREATE_PAGE_DROP_ID,
-  CREATE_PAGE_FIELDS_ORDERED,
-} from './constants'
-export { CreateFieldDrawer } from './CreateFieldDrawer'
