--- frontend/src/features/admin-form-builder/constants.ts
@@ -1,2 +1,0 @@
-export const FIELD_LIST_DROP_ID = 'formFieldList'
-export const PENDING_CREATE_FIELD_ID = 'FIELD-PENDING-CREATION'
