+++ frontend/src/features/admin-form-builder/design/components/BuilderDesignPlaceholder.tsx
@@ -2,9 +2,8 @@ import { useMemo } from 'react'
 import { Box, Flex, Text } from '@chakra-ui/react'
 import { isEmpty } from 'lodash'
 
-import { FIELD_LIST_DROP_ID } from '../../constants'
-
-import { DndPlaceholderProps } from './types'
+import { FIELD_LIST_DROP_ID } from '../constants'
+import { DndPlaceholderProps } from '../types'
 
 export interface BuilderDesignPlaceholderProps {
   placeholderProps: DndPlaceholderProps
