+++ frontend/src/features/admin-form-builder/design/components/BuilderDrawer/CreateFieldDrawer/index.ts
@@ -1,0 +1,1 @@
+export { CreateFieldDrawer } from './CreateFieldDrawer'
