+++ frontend/src/features/admin-form-builder/design/types.ts
@@ -2,6 +2,16 @@ import { FormField, FormFieldDto } from '~shared/types/field'
 
 import { PENDING_CREATE_FIELD_ID } from './constants'
 
+export type DndPlaceholderProps =
+  | {
+      droppableId: string
+      clientHeight: number
+      clientWidth: string
+      clientY: number
+      clientX: number
+    }
+  | Record<string, never>
+
 export type PendingFormField = FormField & {
   _id: typeof PENDING_CREATE_FIELD_ID
 }
